N337 CopySkin from Extra to Tab UI
Removed CopySkin from Extra tab.
Added 2 new buttons to Skinning tab about the CopySkinning from OneSource or with SameName.
