+++ dpAutoRigSystem/Languages/Portugues.json
@@ -5,7 +5,7 @@
     "_date"          : "2011-08-03",
     "_language"      : "Português",
     "_translator"    : "Danilo Pinheiro",
-    "_updated"       : "2023-08-15",
+    "_updated"       : "2023-08-24",
     
     "a000_presentation" : "Esta é a primeira frase a traduzir.\nPor favor, use o campo de texto abaixo para escrever seu texto traduzido.\nVocê pode usar os botões para:\n\nVoltar e revisar sua tradução.\nMesmo para usar o mesmo texto no seu idioma.\nPróximo para aplicar o texto traduzido e seguir em frente.\n\nUse Enter (Return) para uma nova linha.\nAlguns textos não aceitam caracteres especiais.\nOutros têm pontuação ou nada.\n\nVamos começar! Obrigado.",
     
@@ -445,6 +445,10 @@
     "i284_sCallbackAnn"    : "Callback o código Python em init método depois\nde terminar de publicar.\nÉ útil para rodar o próximo processo na pipeline.",
     "i285_alreadySkinned"  : "A geometria já está skinada, desculpe.",
     "i286_displaySkinLog"  : "Mostrar janela de log",
+    "i287_copy"            : "Copiar",
+    "i288_copySkinDesc"    : "Será copiado o skinning do primeiro item selecionado\npara os outros selecionados em seguida.",
+    "i289_sameNameSkinDesc": "Vai copiar o skinning entre meshes de mesmo nome.\nMuito útil para atualizar modelos no rig.",
+    "i290_oneSource"       : "Uma Origem",
 
     "m001_fkLine"               : "Linha Fk",
     "m002_fkLineDesc"           : "Descrição do Modulo Linha Fk:\n\nEsse modulo cria uma cadeia de joints\ncom o numero de joints desejado.\n\nQuando rigado, os controles serão FK (forward kinematics).\n\nEle é util para criar rabos, orelhas, cabelos\nou simples controles de objetos.",
@@ -542,8 +546,8 @@
     "m094_doing"                : "Fazendo",
     "m095_isolate"              : "Isolar",
     "m096_isolateDesc"          : "Esse script vai criar um setup de Isolação.\nPrecisa ter 3 nodes selecionados\nou sera criado usando o Root_Ctrl.",
-    "m097_copySkin"             : "Copiar Skin",
-    "m098_copySkinDesc"         : "Esse script vai copiar o skinning de\numa primeira geometria skinada para\nas outras selecionadas em seguida.",
+    "m097_00000000"             : "NOT USED HERE",
+    "m098_000000000000"         : "NOT USED HERE",
     "m099_cvControlDesc"        : "Essa classe irá criar um shape de curva\npara ser usado como controle.",
     "m100_square"               : "Quadrado",
     "m101_smile"                : "Sorriso",
